+++ CHANGELOG.md
@@ -1,7 +1,37 @@
-# Change Log
-All notable changes to the "atlascode" extension will be documented in this file.
+## What's New In 0.2.0
+### Bug
+* [VSCODE-179](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-179) - Make JQL edit dialog less frustrating
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+### Task
+* [VSCODE-122](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-122) - Quick pick dropdown for selecting Jira project does not show all available projects in a site
+* [VSCODE-135](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-135) - add webview for creating PRs
+* [VSCODE-136](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-136) - update create PR context menu to show new webview
+* [VSCODE-138](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-138) - add custom JQL to jira explorer
+* [VSCODE-142](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-142) - add ability to configure jira auto-refresh time
+* [VSCODE-143](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-143) - add ability to configure BB auto-refresh time
+* [VSCODE-145](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-145) - add info message popup when new PRs are detected
+* [VSCODE-153](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-153) - show related issues for epics on issue view screen
+* [VSCODE-165](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-165) - Use the VS Code clipboard API instead of the library we're using for copy link to PR.
+* [VSCODE-168](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-168) - update 0.1.1 release notes
+* [VSCODE-169](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-169) - Allow multiple JQL queries
+* [VSCODE-170](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-170) - Update SiteJql representation to handle names and enabled/disabled
+* [VSCODE-174](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-174) - Show related Jira issues in pull request webview
+* [VSCODE-177](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-177) - Allow reordering of custom JQL queries
+* [VSCODE-178](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-178) - Update Jira explorer when custom JQL is updated
+* [VSCODE-182](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-182) - Validate source branch by comparing local changes to remote
+* [VSCODE-183](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-183) - Option to push local changes before submitting pull request
+* [VSCODE-185](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-185) - Fix styles on pull request creation webview
+* [VSCODE-186](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-186) - Link to PR tree view after creating pull request
 
-## [Unreleased]
-- Initial release
+
+## What's New In 0.1.1
+### Bug
+* [VSCODE-108](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-108) - Webviews don't work properly in split view.
+
+### Task
+* [VSCODE-48](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-48) - Handle pagination in all Bitbucket apis
+* [VSCODE-131](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-131) - issues in BB explorer
+* [VSCODE-141](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-141) - map BB staging clone urls to BB client base url
+* [VSCODE-161](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-161) - Use diffstat instead of patch to get files changed in a PR
+* [VSCODE-163](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-163) - Integrate JQL autocomplete input
+* [VSCODE-164](https://pi-dev-sandbox.atlassian.net/browse/VSCODE-164) - Persist custom JQL for each site